commit to check status

+++ src/main/java/oopsConcepts/encapsulation/ICICI_Bank.java
@@ -21,7 +21,7 @@ public class ICICI_Bank {
 		accountBalance = accountBalance + 1243567.44;
 		return accountBalance;
 	}
-	public void setAccountBalance(double accountBalance) {
+	public void setAccountBalance(long accountBalance) {
 		this.accountBalance = accountBalance;
 	}
